update total gas fee paid
