+++ src/UI/AnalyticChart/TotalGasFee.svelte
@@ -1,0 +1,162 @@
+<script lang="ts">
+  import { sendMessage } from "webext-bridge";
+  import { wallet, chain } from "~/store";
+  import dayjs from "dayjs";
+  import { formatCurrencyV2 } from "~/utils";
+  import type {
+    AnalyticTotalGasFeeRes,
+    AnalyticTotalGasFeeFormat,
+  } from "~/types/AnalyticTotalGasFeeData";
+
+  import CalendarChart from "~/components/CalendarChart.svelte";
+
+  let selectedWallet: string = "";
+  wallet.subscribe((value) => {
+    selectedWallet = value;
+  });
+
+  let selectedChain: string = "";
+  chain.subscribe((value) => {
+    selectedChain = value;
+  });
+
+  let isLoadingChart = false;
+  let isEmptyDataChart = false;
+  let sum = 0;
+  let option = {
+    tooltip: {
+      extraCssText: "z-index: 9997",
+      formatter: function (params) {
+        return `
+            <div style="display: flex; flex-direction: column; gap: 12px; min-width: 180px;">
+              <div style="font-weight: 500; font-size: 16px; line-height: 19px; color: black;">
+                ${dayjs(params.data[0]).format("DD MMM YYYY")}
+              </div>
+              <div style="display: flex; align-items: centers; justify-content: space-between;">
+                <div style="width: 135px; font-weight: 500; font-size: 14px; line-height: 17px; color: black; display: flex; align-items: centers; gap: 6px;">
+                  <div style="background: #00b580; width: 12px; height: 12px; border-radius: 100%; margin-top: 3px;"></div>
+                  Gas fee paid
+                </div>
+                <div style="display:flex; justify-content: center; align-items: center; gap: 4px; flex: 1; font-weight: 500; font-size: 14px; line-height: 17px; color: #000;">
+                  $${formatCurrencyV2(params.data[1])}
+                </div>
+              </div>
+            </div>`;
+      },
+    },
+    visualMap: {
+      min: 0,
+      max: 1,
+      calculable: true,
+      orient: "horizontal",
+      top: 0,
+      right: 40,
+      inRange: {
+        color: ["#00A878"],
+        opacity: [0, 1],
+      },
+      controller: {
+        inRange: {
+          opacity: [0, 1],
+        },
+        outOfRange: {
+          color: "#f4f5f8",
+        },
+      },
+    },
+    calendar: {
+      top: 80,
+      left: 60,
+      right: 60,
+      cellSize: ["auto", "auto"],
+      range: dayjs(new Date()).format("YYYY"),
+      itemStyle: {
+        borderWidth: 0.5,
+      },
+      yearLabel: { show: false },
+    },
+    dayLabel: {
+      nameMap: ["Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"],
+    },
+    series: {
+      type: "heatmap",
+      coordinateSystem: "calendar",
+      data: [dayjs(new Date()).format("YYYY-MM-DD"), 1],
+    },
+  };
+
+  const getTotalGasFee = async () => {
+    isLoadingChart = true;
+    try {
+      const response: AnalyticTotalGasFeeRes = await sendMessage(
+        "getTotalGasFee",
+        {
+          address: selectedWallet,
+          chain: selectedChain,
+        }
+      );
+
+      if (response?.result && response?.result.length !== 0) {
+        const maxHistorical = response?.result.reduce((prev, current) =>
+          prev.value > current.value ? prev : current
+        );
+
+        sum = response?.result.reduce((prev, item) => prev + item.value, 0);
+
+        const formatData: AnalyticTotalGasFeeFormat = response?.result.map(
+          (item) => {
+            return [dayjs(item.date * 1000).format("YYYY-MM-DD"), item.value];
+          }
+        );
+
+        option = {
+          ...option,
+          visualMap: {
+            ...option.visualMap,
+            max: maxHistorical.value,
+          },
+          calendar: {
+            ...option.calendar,
+            range: dayjs(maxHistorical.date * 1000).format("YYYY"),
+          },
+          series: {
+            ...option.series,
+            data: formatData,
+          },
+        };
+        isLoadingChart = false;
+      } else {
+        isLoadingChart = false;
+        isEmptyDataChart = true;
+      }
+    } catch (e) {
+      console.log("error: ", e);
+      isLoadingChart = false;
+      isEmptyDataChart = true;
+    }
+  };
+
+  $: {
+    if (selectedWallet || selectedChain) {
+      isLoadingChart = false;
+      isEmptyDataChart = false;
+      if (selectedWallet.length !== 0 && selectedChain.length !== 0) {
+        getTotalGasFee();
+      }
+    }
+  }
+</script>
+
+<div class="border border-[#0000001a] rounded-[20px] py-10">
+  <CalendarChart
+    {option}
+    {isEmptyDataChart}
+    {isLoadingChart}
+    title="The total gas fee paid"
+    tooltipTitle=""
+    id="TotalGasFeePaid"
+    {sum}
+  />
+</div>
+
+<style windi:preflights:global windi:safelist:global></style>
