+++ src/entries/background/main.ts
@@ -56,7 +56,20 @@ onMessage<ICoinListInput, any>("coinList", async ({ data: { limit } }) => {
 
 onMessage<IAddressInput, any>("getSectorGrowth", async ({ data: { address, chain } }) => {
   try {
-    return nimbus.get(`/analysis/${address}/sector-growth?chain=${chain}`).then((response) => {
+    return nimbus.get(`/analysis/${address}/sector-growth?chain=${chain}&fromDate=${""}&toDate=${""}`).then((response) => {
+      return {
+        result: response.data,
+        address: address
+      }
+    });
+  } catch (error) {
+    return {};
+  }
+});
+
+onMessage<IAddressInput, any>("getTotalGasFee", async ({ data: { address, chain } }) => {
+  try {
+    return nimbus.get(`/analysis/${address}/gas-used?chain=${chain}&fromDate=${""}&toDate=${""}`).then((response) => {
       return {
         result: response.data,
         address: address
