+++ src/types/AnalyticHistoricalData.ts
@@ -4,3 +4,5 @@ export type AnalyticHistorical = {
   count: number,
   date: string
 }
+
+export type AnalyticHistoricalFormat = any[]
